Rename poc-cluster environment to have "-running" suffix
Kubernetes is not starting in the conjur-k8s-secretless scenario:

    https://www.katacoda.com/conjur/scenarios/conjur-k8s-secretless

This is because /usr/bin/launch.sh doesn't try to start Kubernetes; it
just waits for the startup and configuration to be complete.

The regular multi-node tutorial
(https://www.katacoda.com/courses/kubernetes/getting-started-with-kubeadm)
starts off with two code snippets in step 1 to start the cluster and
configure kubectl.  Currently, Katacoda will run these commands
automatically if the cluster name includes "-running".

Rename the poc-cluster to poc-cluster-running.